+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 0.0.6
 - Added GitHub issue template
 - Added more i18n
-- Added auto play feature
+- Added auto-play feature
+- Added lodash pacakge
 - Updated audio element in slide show
 - Update toggle hook (useState -> useBoolean hook in Chakra UI)